[#858] Move izonereport tests into their own test case file
